+++ lib/modules/main/domain_layer/use_cases/done_section_use_case.dart
@@ -6,7 +6,7 @@ class DoneSectionUseCase {
   DoneSectionUseCase(this.baseMainRepository);
   Future<Either<Exception, void>> done({
     required String courseName,
-    required double progress,
+    required int progress,
     required int done,
   }) {
     return baseMainRepository.doneSection(
